rf: move `DandiModelLinkmlValidator`
From `cli/tools.py` to `tools.py`
